codebox-agent/pyproject.toml
[project]
name = "codebox-agent"
version = "0.1.0"
description = "Portable AI coding agent — LangGraph + deepagents foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents",
    "langchain-openai",
    "langchain-openrouter",
    "openrouter>=0.7,<0.8",
    "langgraph-checkpoint-sqlite",
    "pydantic>=2.0",
    "python-dotenv",
    "tavily-python",
    "markdownify",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codebox_agent"]
